// backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const Role = {
  ATTENDEE: 'ATTENDEE',
  ORGANIZER: 'ORGANIZER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PaymentStatus = {
  WAITING_FOR_PAYMENT: 'WAITING_FOR_PAYMENT',
  WAITING_FOR_ADMIN_CONFIRMATION: 'WAITING_FOR_ADMIN_CONFIRMATION',
  DONE: 'DONE',
  REJECTED: 'REJECTED',
  EXPIRED: 'EXPIRED',
  CANCELED: 'CANCELED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const EventCategory = {
  MUSIC: 'MUSIC',
  NIGHTLIFE: 'NIGHTLIFE',
  WORKSHOP: 'WORKSHOP',
  CONFERENCE: 'CONFERENCE',
  EXHIBITION: 'EXHIBITION',
  SPORTS: 'SPORTS',
  FESTIVAL: 'FESTIVAL',
  COMMUNITY: 'COMMUNITY',
  OTHER: 'OTHER'
} as const

export type EventCategory = (typeof EventCategory)[keyof typeof EventCategory]